feat: introduced command validators

+++ HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -11,6 +11,11 @@ namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeave
         public DeleteLeaveTypeCommandHandler(ILeaveTypeRepository leaveTypeRepository) => _leaveTypeRepository = leaveTypeRepository;
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
+            var validator = new DeleteLeaveTypeCommandValidator(_leaveTypeRepository);
+            var validatorResult = await validator.ValidateAsync(request);
+            if (validatorResult.Errors.Any())
+                throw new BadRequestException("Invalid Leavetype", validatorResult);
+
             var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id) ?? throw new NotFoundException(nameof(Domain.LeaveType), request.Id);
             await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
             return Unit.Value;
